Migrate Stage 5 checks to automated_checks.py, remove deprecated lead_verification.py
- Migrated all Stage 5 verification logic (role, region, industry) to automated_checks.py
- Implemented early exit logic for Stage 5 checks
- Added fuzzy matching and DDG search improvements
- Fixed LinkedIn verification with URL exact match override
- Removed deprecated lead_verification.py
